Fix tests and work in review
